Lesson 05_03: Borrow Javascript from Twenty Seventeen to enable dropdown menus
Scrap underscores navigation.js in favor of the more accessible nav setup from twentyseventeen. This was all well and good except that the sub-menus didn't expand. There were 2 issues, one was styling. I overwrote my whole /sass directory with the lesson files, then the toggles at least showed up. But the issue with the script was that the initial state of the active sub-menu was looking for .sub-menu, which is not used. So I changed that to .children and everything works!