Lesson 06_03: Modify the output of post meta data
We reversed the order of author and date, and move the comment meta and edit link to the top meta, out of the footer. My version of underscores had separate functions for both the author and the publish date, while the example was using a single function. So instead of choosing which function to combine the comments and edit link with, I made them their own standalone functions in template-tags.php as well. I see ways that this could be very useful, it will be much easier from a templating standpoint to be able to move them all individually at will.