Lesson 07_03: Building the structure for displaying the sidebar
We created a new content template for single posts, and called it from single.php. We then moved the sidebar from single.php into the new content-single template. We moved post navigation and comments into content-single, and then wrapped a section around the entry content, post nav and comments. This gives us the structure to create the desired layout