Lesson 07_04: Use Flexbox to create a responsive layout
Well FUCK I definitely overwrote my changes like 3 times. ANYWAY, we used really basic flex, and some width variables, in order to create the structure of a full width post header, section of content, comments, post nav to the left, and sidebar to the right. It's all constrained in a max width so it's no longer fluid to the whole site. I'm going to review git on my lunch! FML