Lesson 07_05: Customize the single post template for a no-sidebar layout
I had to re-add all of my individual post meta functions, at some point I overwrote with his single function call. Rethinking that strategy, but can deal with it later. We removed all the widgets to get the no-sidebar class on the body. We conditionally displayed the meta in the post header if sidebar is active, and in the post content if sidebar is not active. We added 2 new body classes conditionally for sidebar is not active as well, one to wrap both post content elements, and one to just wrap the post content body (entry and footer), so we can use flex to float them next to each other